Split Light Environment editor owner
