+++ scripts/quality-baseline.json
@@ -7,6 +7,6 @@
   "viewRuntimeBridgeLookups": 0,
   "labStateAppFiles": 0,
   "labStateTestFiles": 0,
-  "largeJsFilesOver800Lines": 19,
+  "largeJsFilesOver800Lines": 18,
   "maxJsFileLines": 1168
 }
