+++ tests/test-correctness-phase2.js
@@ -161,6 +161,7 @@ const pwaAppShellAssets = [
   '/js/light-device-session-engine.js',
   '/js/light-device-setup-modal.js',
   '/js/light-env.js',
+  '/js/light-env-editor.js',
   '/js/light-env-screen-ui.js',
   '/js/light-env-audits.js',
   '/js/light-env-evening.js',
