+++ tests/test-quality-guardrails.js
@@ -242,6 +242,7 @@ const domainUiCheckJsModules = [
   'js/lab-context.js',
   'js/light-conditions-now.js',
   'js/light-env.js',
+  'js/light-env-editor.js',
   'js/marker-detail-modal.js',
   'js/marker-detail-runtime.js',
   'js/pdf-import-review.js',
@@ -376,6 +377,7 @@ const lightWorkflowCheckJsModules = [
   'js/light-env-actions.js',
   'js/light-env-ai-analysis.js',
   'js/light-env-audits.js',
+  'js/light-env-editor.js',
   'js/light-env-evening.js',
   'js/light-env-model.js',
   'js/light-env-screen-ui.js',
