+++ tsconfig.json
@@ -173,6 +173,7 @@
     "js/light-env-ai-analysis.js",
     "js/light-env-audits.js",
     "js/light-env-evening.js",
+    "js/light-env-editor.js",
     "js/light-env-model.js",
     "js/light-env-store.js",
     "js/light-env.js",
